Get weather package option

Retrieves the value of a specified option from the weather package configuration. This function provides access to package-level settings and preferences.

Usage

weather_get_option(name)

Arguments

name

A character string specifying the name of the option to retrieve.

Value

The value of the specified option, or NULL if the option does not exist.

Examples

# Get the default frost threshold
weather_get_option("extreme.frost_threshold")
#> [1] 0